refactor: reorganize waitready code per review
Move WaitForCephReady to dedicated waitready.go (not monitor-specific),
move OSD wait functions (WaitForOSDsReady, getRequiredOSDCount,
getUpOSDCount, getDefaultPoolSize) to osd.go where they belong.

Split install_microceph into install/bootstrap phases and combine
pre/post bootstrap waitready tests into a single test_waitready function
that exercises both scenarios without a purge/reinstall cycle.
