+++ microceph/ceph/monitor.go
@@ -1,13 +1,9 @@
 package ceph
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/canonical/microceph/microceph/common"
 	"github.com/canonical/microceph/microceph/interfaces"
@@ -128,95 +124,4 @@ func getActiveMons() ([]string, error) {
 	}
 
 	return activeMons, nil
-}
-
-// WaitForCephReady polls "ceph -s" until it succeeds, indicating that
-// the Ceph can accept commands.
-// It retries every second until success or the context is cancelled/expired.
-func WaitForCephReady(ctx context.Context) error {
-	for {
-		_, err := cephRunContext(ctx, "-s")
-		if err == nil {
-			return nil
-		}
-
-		logger.Debugf("Ceph not ready yet: %v", err)
-
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("timed out waiting for Ceph to become ready: %w", ctx.Err())
-		case <-time.After(time.Second):
-		}
-	}
-}
-
-// WaitForOSDsReady polls until enough OSDs are up to satisfy pool replication
-// requirements. The required count is max(pool.Size) across all pools, falling
-// back to osd_pool_default_size if no pools exist.
-func WaitForOSDsReady(ctx context.Context) error {
-	for {
-		required, err := getRequiredOSDCount(ctx)
-		if err != nil {
-			logger.Debugf("Failed to determine required OSD count: %v", err)
-		} else {
-			count, err := getUpOSDCount(ctx)
-			if err == nil && count >= required {
-				return nil
-			}
-		}
-
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("timed out waiting for OSDs to be ready: %w", ctx.Err())
-		case <-time.After(time.Second):
-		}
-	}
-}
-
-// getRequiredOSDCount returns the maximum pool size across all pools, or the
-// default pool size if no pools exist.
-func getRequiredOSDCount(ctx context.Context) (int64, error) {
-	pools, err := GetOSDPools(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("failed to get pools: %w", err)
-	}
-
-	if len(pools) == 0 {
-		return getDefaultPoolSize(ctx)
-	}
-
-	var maxSize int64
-	for _, pool := range pools {
-		if pool.Size > maxSize {
-			maxSize = pool.Size
-		}
-	}
-
-	return maxSize, nil
-}
-
-// getDefaultPoolSize reads the osd_pool_default_size configuration value.
-func getDefaultPoolSize(ctx context.Context) (int64, error) {
-	output, err := cephRunContext(ctx, "config", "get", "mon", "osd_pool_default_size")
-	if err != nil {
-		return 0, fmt.Errorf("failed to get osd_pool_default_size: %w", err)
-	}
-
-	size, err := strconv.ParseInt(strings.TrimSpace(output), 10, 64)
-	if err != nil {
-		return 0, fmt.Errorf("failed to parse osd_pool_default_size %q: %w", output, err)
-	}
-
-	return size, nil
-}
-
-// getUpOSDCount returns the number of OSDs currently in the up state.
-func getUpOSDCount(ctx context.Context) (int64, error) {
-	output, err := cephRunContext(ctx, "osd", "dump", "-f", "json-pretty")
-	if err != nil {
-		return 0, fmt.Errorf("failed to get OSD dump: %w", err)
-	}
-
-	upOsds := gjson.Get(output, "osds.#(up==1)#.uuid") // select all uuids where the up field equals 1
-	return int64(len(upOsds.Array())), nil
 }
