+++ microceph/ceph/osd_test.go
@@ -2,12 +2,14 @@ package ceph
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"github.com/canonical/microceph/microceph/common"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/canonical/lxd/shared/api"
+	"github.com/canonical/microceph/microceph/common"
 	"github.com/canonical/microceph/microceph/tests"
 	"github.com/spf13/afero"
 
@@ -996,4 +998,140 @@ func (s *osdSuite) TestIsPristineDisk() {
 	assert.Error(s.T(), err)
 	assert.False(s.T(), isPristine)
 	assert.Contains(s.T(), err.Error(), "permission denied")
+}
+
+// osdDumpJSON generates a JSON OSD dump with the given number of up OSDs.
+func osdDumpJSON(upCount int) string {
+	return osdDumpJSONMixed(upCount, 0)
+}
+
+// osdDumpJSONMixed generates a JSON OSD dump with a mix of up and down OSDs.
+func osdDumpJSONMixed(upCount, downCount int) string {
+	osds := "["
+	for i := 0; i < upCount; i++ {
+		if i > 0 {
+			osds += ","
+		}
+		osds += fmt.Sprintf(`{"uuid":"osd-%d","up":1}`, i)
+	}
+	for i := 0; i < downCount; i++ {
+		if upCount > 0 || i > 0 {
+			osds += ","
+		}
+		osds += fmt.Sprintf(`{"uuid":"osd-down-%d","up":0}`, i)
+	}
+	osds += "]"
+	return fmt.Sprintf(`{"osds":%s}`, osds)
+}
+
+func (s *osdSuite) TestWaitForOSDsReadyImmediateSuccess() {
+	r := mocks.NewRunner(s.T())
+
+	// 1 pool with size=1.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "ls", "--format", "json").
+		Return(`["mypool"]`, nil).Once()
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "get", "mypool", "all", "--format", "json").
+		Return(`{"pool":"mypool","pool_id":1,"size":1,"min_size":1,"crush_rule":""}`, nil).Once()
+	// Found 1 OSD.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSON(1), nil).Once()
+
+	common.ProcessExec = r
+
+	// Should succeed: required 1 replica, found 1 OSD.
+	err := WaitForOSDsReady(context.Background())
+	assert.NoError(s.T(), err)
+	r.AssertExpectations(s.T())
+}
+
+func (s *osdSuite) TestWaitForOSDsReadyRetriesThenSucceeds() {
+	r := mocks.NewRunner(s.T())
+
+	// Pool with size=3 (called each iteration).
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "ls", "--format", "json").
+		Return(`["rbd"]`, nil)
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "get", "rbd", "all", "--format", "json").
+		Return(`{"pool":"rbd","pool_id":1,"size":3,"min_size":2,"crush_rule":""}`, nil)
+	// First poll: only 2 OSDs up.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSON(2), nil).Once()
+	// Second poll: 3 OSDs up.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSON(3), nil).Once()
+
+	common.ProcessExec = r
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	err := WaitForOSDsReady(ctx)
+	assert.NoError(s.T(), err)
+	r.AssertExpectations(s.T())
+}
+
+func (s *osdSuite) TestWaitForOSDsReadyWithDownOSDs() {
+	r := mocks.NewRunner(s.T())
+
+	// Pool with size=3 (called each iteration).
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "ls", "--format", "json").
+		Return(`["rbd"]`, nil)
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "get", "rbd", "all", "--format", "json").
+		Return(`{"pool":"rbd","pool_id":1,"size":3,"min_size":2,"crush_rule":""}`, nil)
+	// First poll: 2 up, 2 down — not enough.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSONMixed(2, 2), nil).Once()
+	// Second poll: 3 up, 1 down — sufficient.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSONMixed(3, 1), nil).Once()
+
+	common.ProcessExec = r
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	err := WaitForOSDsReady(ctx)
+	assert.NoError(s.T(), err)
+	r.AssertExpectations(s.T())
+}
+
+func (s *osdSuite) TestWaitForOSDsReadyFallbackToDefault() {
+	r := mocks.NewRunner(s.T())
+
+	// No pools.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "ls", "--format", "json").
+		Return(`[]`, nil).Once()
+	// Default size = 3.
+	r.On("RunCommandContext", mock.Anything, "ceph", "config", "get", "mon", "osd_pool_default_size").
+		Return("3\n", nil).Once()
+	// 3 OSDs up.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return(osdDumpJSON(3), nil).Once()
+
+	common.ProcessExec = r
+
+	err := WaitForOSDsReady(context.Background())
+	assert.NoError(s.T(), err)
+	r.AssertExpectations(s.T())
+}
+
+func (s *osdSuite) TestWaitForOSDsReadyTimeout() {
+	r := mocks.NewRunner(s.T())
+
+	// 1 pool with size=1 (called each iteration).
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "ls", "--format", "json").
+		Return(`["mypool"]`, nil)
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "pool", "get", "mypool", "all", "--format", "json").
+		Return(`{"pool":"mypool","pool_id":1,"size":1,"min_size":1,"crush_rule":""}`, nil)
+	// OSD dump always fails.
+	r.On("RunCommandContext", mock.Anything, "ceph", "osd", "dump", "-f", "json-pretty").
+		Return("", errors.New("connection refused"))
+
+	common.ProcessExec = r
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Millisecond)
+	defer cancel()
+
+	err := WaitForOSDsReady(ctx)
+	assert.Error(s.T(), err)
+	assert.ErrorIs(s.T(), err, context.DeadlineExceeded)
 }
