+++ tests/scripts/actionutils.sh
@@ -24,7 +24,6 @@ function setup_lxd() {
 }
 
 function install_microceph() {
-    local mon_ip="${1-}"
     # Install locally built microceph snap and connect interfaces
     sudo snap install --dangerous ~/microceph_*.snap
     sudo snap connect microceph:block-devices
@@ -35,6 +34,10 @@ function install_microceph() {
     sudo snap connect microceph:microceph-support
     sudo snap connect microceph:network-bind
     sudo snap connect microceph:process-control
+}
+
+function bootstrap_microceph() {
+    local mon_ip="${1-}"
 
     if [ -n "${mon_ip}" ]; then
         sudo microceph cluster bootstrap --mon-ip "${mon_ip}"
@@ -48,6 +51,12 @@ function install_microceph() {
     sleep 30
     sudo microceph.ceph status
     sudo microceph.ceph health
+}
+
+function install_and_bootstrap_microceph() {
+    local mon_ip="${1-}"
+    install_microceph
+    bootstrap_microceph "${mon_ip}"
 }
 
 function install_tools() {
@@ -1524,7 +1533,7 @@ function verify_pristine_check() {
 
     # Create some dirty disks
     create_loop_devices
-    install_microceph
+    install_and_bootstrap_microceph
     sudo microceph disk add /dev/sdia --wipe
     sudo microceph disk add /dev/sdib --wipe
     sudo microceph disk add /dev/sdic --wipe
@@ -1533,7 +1542,7 @@ function verify_pristine_check() {
     
     # Remove the snap to leave dirty disks then reinstall
     sudo snap remove microceph --purge
-    install_microceph
+    install_and_bootstrap_microceph
 
     # Levels of dirtiness
     # - leave sdia completely dirty
@@ -1583,20 +1592,22 @@ function verify_mount_check() {
     fi
 }
 
-function test_waitready_before_bootstrap() {
-    # waitready should fail before any cluster is bootstrapped.
+function test_waitready() {
+    set -eux
+
+    install_microceph
+
+    # Pre-bootstrap: should fail
     if sudo microceph waitready --timeout 5 2>/dev/null; then
-        echo "ERROR: waitready succeeded before bootstrap, expected failure"
+        echo "ERROR: waitready succeeded before bootstrap"
         exit 1
     fi
     echo "waitready correctly failed before bootstrap"
-}
 
-function test_waitready_after_bootstrap() {
-    set -e
-    # After bootstrap, waitready should succeed within the timeout.
+    bootstrap_microceph
+
+    # Post-bootstrap: should succeed
     sudo microceph waitready --timeout 30
-    sudo microceph disk list
     echo "waitready succeeded after bootstrap"
 }
